+++ src/main/scala/com/crealytics/spark/excel/DefaultSource.scala
@@ -24,6 +24,7 @@ class DefaultSource extends RelationProvider with SchemaRelationProvider with Cr
     ExcelRelation(
       header = checkParameter(parameters, "header").toBoolean,
       treatEmptyValuesAsNulls = parameters.get("treatEmptyValuesAsNulls").fold(false)(_.toBoolean),
+      usePlainNumberFormat = parameters.get("usePlainNumberFormat").fold(false)(_.toBoolean),
       userSchema = Option(schema),
       inferSheetSchema = parameters.get("inferSchema").fold(false)(_.toBoolean),
       addColorColumns = parameters.get("addColorColumns").fold(false)(_.toBoolean),